Title: 씨 오 더블유, Time: 133ms, Memory: 13MB, Status: Passed - CodeTree

+++ 241117/씨 오 더블유/c-o-w.java
@@ -29,7 +29,7 @@ public class Main {
             }
         }
 
-        int cnt = 0;
+        long cnt = 0;
         for (int i = 0; i < n; i++) {
             if (str.charAt(i) == 'C') {
                 cnt += owArray[i];
